requirements.txt
tensorflow
numpy
Pillow
tk
